+++ src/Stmt.cc
@@ -1331,7 +1331,7 @@ ValPtr ForStmt::DoExec(Frame* f, Val* v, StmtFlowType& flow)
 		for ( const auto& lve : *loop_vals )
 			{
 			auto k = lve.GetHashKey();
-			auto* current_tev = lve.GetValue<TableEntryVal*>();
+			auto* current_tev = lve.value;
 			auto ind_lv = tv->RecreateIndex(*k);
 
 			if ( value_var )
