+++ src/Val.h
@@ -1044,7 +1044,7 @@ protected:
 	detail::ExprPtr expire_time;
 	detail::ExprPtr expire_func;
 	TableValTimer* timer;
-	RobustDictIterator* expire_iterator;
+	RobustDictIterator<TableEntryVal>* expire_iterator;
 	detail::PrefixTable* subnets;
 	ValPtr def_val;
 	detail::ExprPtr change_func;
