+++ src/broker/Data.cc
@@ -924,8 +924,6 @@ broker::expected<broker::data> val_to_data(const Val* v)
 			for ( const auto& te : *table )
 				{
 				auto hk = te.GetHashKey();
-				auto* entry = te.GetValue<TableEntryVal*>();
-
 				auto vl = table_val->RecreateIndex(*hk);
 
 				broker::vector composite_key;
@@ -952,7 +950,7 @@ broker::expected<broker::data> val_to_data(const Val* v)
 					get<broker::set>(rval).emplace(move(key));
 				else
 					{
-					auto val = val_to_data(entry->GetVal().get());
+					auto val = val_to_data(te.value->GetVal().get());
 
 					if ( ! val )
 						return broker::ec::invalid_data;
