+++ src/file_analysis/AnalyzerSet.h
@@ -96,7 +96,8 @@ public:
 	void DrainModifications();
 
 	// Iterator support
-	using iterator = zeek::DictIterator;
+	using iterator = zeek::DictIterator<file_analysis::Analyzer>;
+	;
 	using const_iterator = const iterator;
 	using reverse_iterator = std::reverse_iterator<iterator>;
 	using const_reverse_iterator = std::reverse_iterator<const_iterator>;
